Project - 14

+++ controllers/cards.js
@@ -1,17 +1,18 @@
 const Card = require('../models/card');
 const NotFoundError = require('../errors/NotFoundError');
-const { handleError } = require('../errors/handleError');
+const ForbiddenError = require('../errors/ForbiddenError');
+const ValidationError = require('../errors/ValidationError');
 
-module.exports.getCards = async (req, res) => {
+module.exports.getCards = async (req, res, next) => {
   try {
     const cards = await Card.find({}).populate(['owner', 'likes']);
     res.send(cards);
   } catch (err) {
-    handleError({ err, res });
+    next(err);
   }
 };
 
-module.exports.createCard = async (req, res) => {
+module.exports.createCard = async (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   try {
@@ -19,21 +20,29 @@ module.exports.createCard = async (req, res) => {
     await card.populate(['owner', 'likes']);
     res.send(card);
   } catch (err) {
-    handleError({ err, res });
+    if (err.name === 'ValidationError' || err.name === 'CastError') {
+      next(new ValidationError('Переданы некорректные данные'));
+    } else next(err);
   }
 };
 
-module.exports.deleteCard = async (req, res) => {
+module.exports.deleteCard = async (req, res, next) => {
   try {
-    const card = await Card.findByIdAndDelete(req.params.cardId);
-    if (card) res.send({ message: 'Карточка удалена' });
-    else throw new NotFoundError('Карточка не найдена');
+    const card = await Card.findById(req.params.cardId);
+    if (card) {
+      const isOwner = req.user._id === card.owner.toString();
+      if (!isOwner) throw new ForbiddenError('Вы не являетесь владельцем карточки');
+      await Card.findByIdAndDelete(req.params.cardId);
+      res.send({ message: 'Карточка удалена' });
+    } else throw new NotFoundError('Карточка не найдена');
   } catch (err) {
-    handleError({ err, res });
+    if (err.name === 'ValidationError' || err.name === 'CastError') {
+      next(new ValidationError('Переданы некорректные данные'));
+    } else next(err);
   }
 };
 
-module.exports.putLike = async (req, res) => {
+module.exports.putLike = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.cardId);
     if (card) {
@@ -46,11 +55,13 @@ module.exports.putLike = async (req, res) => {
       res.send(updatedCard);
     } else throw new NotFoundError('Карточка не найдена');
   } catch (err) {
-    handleError({ err, res });
+    if (err.name === 'ValidationError' || err.name === 'CastError') {
+      next(new ValidationError('Переданы некорректные данные'));
+    } else next(err);
   }
 };
 
-module.exports.deleteLike = async (req, res) => {
+module.exports.deleteLike = async (req, res, next) => {
   try {
     const card = await Card.findById(req.params.cardId);
     if (card) {
@@ -63,6 +74,8 @@ module.exports.deleteLike = async (req, res) => {
       res.send(updatedCard);
     } else throw new NotFoundError('Карточка не найдена');
   } catch (err) {
-    handleError({ err, res });
+    if (err.name === 'ValidationError' || err.name === 'CastError') {
+      next(new ValidationError('Переданы некорректные данные'));
+    } else next(err);
   }
 };
